requirements.txt
# Core ML/DL frameworks
torch>=2.6.0
torchvision>=0.17.0
transformers>=4.30.2
scikit-learn>=1.3.0
scipy>=1.11.1
numpy>=1.24.3
pandas>=2.0.3

# OCR engines
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pytesseract>=0.3.10
pillow>=10.0.0

# Computer Vision
opencv-python>=4.8.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.1
python-multipart>=0.0.6
pydantic>=2.0.3

# MLOps
mlflow>=2.5.0
dvc>=3.15.0

# Data Processing
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.7.0

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2